extract token create/decode logic from auth class
