+++ project/src/database/repositories/user.py
@@ -2,7 +2,7 @@ import sqlalchemy
 from sqlalchemy.orm import Session
 
 from src.database.models import user
-from src.business_logic.authorization import Authorization
+from src.business_logic.classes.authentication import Authentication
 from .repository import AbstractRepository
 
 
@@ -23,7 +23,7 @@ class UserRepository(AbstractRepository):
         name: str | None,
         photo: str | None = None,
     ) -> None:
-        hashed_password = Authorization().get_hashed_password(password)
+        hashed_password = Authentication().get_hashed_password(password)
         query = sqlalchemy.insert(user.Profile).values(
             email=email, password=hashed_password, name=name, photo=photo
         )
